Committing a physical inventory seems to be working
Switched to using the view to collect all the proposed corrections.  The
app code gets the view data as JSON, and then makes inventory correction
docs out of it.

The new view should be able to be used to query the current inventory levels when you
limit the results where the first key is 1 (is-permanent)