Fix a bug where the OrderWidget's callbacks got item-updated and customer-updated could misfire
Sammy activities have a way to bind a callback but no way to unbind it.
There's also no clean way to know when the user leaves the rceive/sale page.

These two things combined to make a memory leak where the OrderWidget binds
to the 'user-updated' and 'customer-updated' signals, but can't stop listening
when they go to another page.  If they then go to the item lister and add or
edit an item, the signal still fires.  Additionally, each time the user went
to the sale/receive page, another identical listener would get registered.

This fix ensures that only one listener will ever get installed.

It also has an ugly fix so the activity tells the OrderWidget to stop
processing the signal when it comes through.  Setting up A before() handler
to turn off processing didn't work, because it would incorrectly fire after
the item form gets submitted and the history goes back to the sale/receive
form