basic shipment summary report works when called directly
Its not hooked up to the interface yet

Shows one block for each shipment in the specified date range.

The line items are only the items included in that shipment, even if there are
other shipments for this order.

The "Order total", "tax" and "shipping charge" are for the entire order.