Refactor item validations again
There's now a list of functions that all need to pass for an object to
be valid.

The validate() method runs through them.  An exception thrown by any of the
functions works as normal.

A new method validateAll() ensures all the validations are called, even if
one or more of them throw exceptions.