Fix the "Reduce must shrink more rapidly" bug
The reduce for the most popular item views does reduce to a single object in
each pass, but that object keeps growing in size.  After there are 30 or 40
sold items, you start getting the above error and nothing works.

Instead of doing the aggregation in the reduce step, do the same work in the
list handler.

Another possible solution would be to set up a temp DB and do a filtered
replication to copy on the sale orders in the requested date range and order
source.  Then have a view in the temp DB that emits keys with the item counts
and values are the item data.  That view can be queried either ascending or
descending, and have a limit.  Finally, remove the temp DB.