Refactor to the dialog modal code
get rid of the old dialogModal() and make newdialogModal() the default

Add markup to the buttons so we can get rid of the submit() handler in
the code.  The markup + Bootstrap's modal stuff makes all the magic happen.

Get rid of the <form> around the buttons.  It makes it look better