Allow the user to specify an ID for warehouse transfers
Fills the place of the order number for other kinds of orders.  If you leave
the input blank, it uses the old behavior of assigning an ID that looks like
"xfer-name1-name2-date

Amazon in particular supplies an ID when shipping items to them.  This is
where you'd put that info into the system