Fix a bug in the order widget
Should finally fix the bug where if you scan in an unknown item, identify the
item, then later scan the item's barcode, it would add additional, duplicate
line-items to the order.  It would also happen if you scanned the sku, then
later scanned the barcode.

This should fix all the combinations of scanning the original (unknown) identifier,
the barcode and the sku, all in the same session.